feat: remove compass-rails dependency
compass-rails is a particularly lacking gem. it hasn't worked with sprockets for the past 2 versions of Rails. fortunately, we aren't using much of it. this commit removes all compass-rails dependencies.

NOTE: this is not an unopnionated update. in removing compass-rails, i did not make an effort to support legacy browser but only our IE9 target. normalize.css has also been updated from 1.0.1 to 3.0.2, removing any vendor support we were getting for legacy browsers. I did make an effort to set reasonable fallbacks where it seemed neccessary.

* gem: remove compass-rails
* gem: remove sprockets version lock

* refactor: sensibly reorder `interfaces/interfaces.css` import list
* refactor: de-compass alerts
* refactor: de-compass badges
* refactor: de-compass buttons
* refactor: de-compass default_layout
* refactor: de-compass forms
* refactor: de-compass lists
* refactor: de-compass modals
* refactor: de-compass panes
* refactor: de-compass tab-list
* refactor: de-compass tables
* refactor: de-compass tooltips
* refactor: de-compass top_bar
* refactor: de-compass widgets
* refactor: de-compass mq/med-screen
* refactor: de-compass mq/small-screen-max
* refactor: de-compass mq/xlarge

* fix: add `base-line-height` definition from compass/vertical-rhythm
* fix: add 100%-height rules from bastardized normalize.sass
* fix: add top_bar ul rules from bastardized normalize.sass
* fix: add `pointable` import to dropdown-menu-item

* remove: shared_compass_dependencies
* remove: bastardized normalize.sass (v1.0.1)

* add: normalize.css (v3.0.2)