fix(Services): don't render mobile-topbar for services
The route section of the new Topbar yields to content given by the Host app. It duplicates the content into 'topbar' and 'mobile-topbar'.

Services doesn't yet support the mobile breakpoint and the content it passed in has an id. So, yah, this sucks. Only in Services, I need to not render the 'mobile-topbar'.

Eventually, 'concuren_plans' section will need to be reworked. That day is not today.