feat(MobileTopbarRouteList): allow button styling
Issue:

Applications like check-ins have an additional action button on the
topbar. The Interfacse MobileTopbar doesn't have a method of adding
styles for thos buttons, in the mobile layout.

Solution:

Add a class that will get observed (form the desktop version) and placed
with a `mobile-` prefixed equivalent for mobile.

Outcome/side effects:

* It's gross (but easy to find when i want to change it)
* There's strange issue I wasn't able to fully explore with regard to
  turbolinks and the way I'm generating the mobile routes. I'm bypassing
  it by using `data-no-turbolink`. I'm not sure that's a long-term
  solution for big route-changes like this. But, I'll cross that bridge
  when it makes a Trello card for me.