fix(MobileTopbar): to render default 'menu' when no active route is rendered
Issue:

`is-active` is implemented app-side. It's not always the case that an active route results in `is-active` being added to a primary route. This can be a result of error or a route being outside of the main heirarchy.

Solution:

Check to see if an active route was found. If so, return it. If not, return "menu"