modify(MobileTopbar): show all major-routes all the time
Update MobileTopbar menu to always show all major routes.

Before this commit, MobileTopbar excluded the current major-route.

This is because I thought it'd be weird to see the currount major route both as the the button and a menu option.

Now, all major modes are shown, regardless of which route is current. This allows users that are drilled deep into a rout access to the topmost point.

Additionally, the current major-route is highlighted to show that reiterate which part of the navigation tree you're in.

In the future, it would be nice for MobileTopbarRouteButton to take present titles with more granularity than major-route.