style(app-badge): remove dark background color
Before this commit, the app-badge had a dark background color. There was a source of confusion for some, who saw the the sidebar as visually connected to the app-badge.

Now, the app-badge reveals the topbar color.

Additionally, hover and border styles were added where appropriate. `icons-extended` was created and added to the interfaces manifest. This file should hold additional class-names, used by interfaces, that aren't generated by icomoon.