feat(topbar): add app-settings button
Before this commit, there was no application-wide settings button.

Now, you can opt into a settings cog to the right of the help icon. You opt-in by passing the `setting_path` argement to the `interfaces_header` helper. Like so:

    <%= interfaces_header person: current_person, organization: current_organization, settings_path: settinsg_path do %>
    <% end %>

Additionally, styles were adjusted to support a `topbar_help-action` being an `a`-tag.