feat: add .view construct for creating views
Before this commit, there was no universal construct for creating views. This was completely left up to app-teams and devs.

Now, `.view` is the construct that should be used to wrap all non-modal rails views. This comes with additional classes that help manage layouts.

    .view - general use flex container (column)
    .view_body - the bulk of your view body. non-header|footer content
    .view_content - ensures that all available space is accounted for, after sidebars.
    .item-sidebar - this class makes an aside the same width as the app-badge and aligns it left
    .item-sidebar-list - this extends `list` with some sensible defaults for an item sidebar

    example:

    <div class="view">
      <h1> Header </h1>

      <div class="view_body">
        <aside class="item-sidebar item-aside-list list">
          <a href="#" class="liste-item">Item navigation #1</a>
          <a href="#" class="liste-item">Item navigation #2</a>
        </aside>

        <div class="view_content">
          ...
        </div>
      </div>
    </div>

Headers, footers, and asides don't require any additional structure classes. I care very little what is used, just that when you use it, it works as expected.