chore: remove modernizr-rails
Before this commit, we were using the modernizr-rails gem and loading heaven knows what.

Now, Interfaces uses a custom build of Modernizr v3.3 that only checks for touch events. This is all that interfaces was using it for. It's possible thatother apps were doing deeper checks. Give it a look in your apps.