chore: move stylesheet 'modules' -> 'preset'
Before this commit, all interfaces stylesheets lived in the "modules" directory.

This is because, it originated as a Foundation clone.

Now, I've moved all "modules" into "preset". This is part of a much larger effort to make Interfaces more maintainable.
Presets will become the existing public API of Interfaces.

This may seem like a pendantic change but it enables a few important changes:

The story for how things get added to interfaces has never been particularly well fleshed out.
This is largely do to a lack of "channels" in Interfaces. Using Interfaces you get EVERYTHING,
making it hard to "test-drive" features without signing up to upgrade for every little change.

If you subscribe only co "presets", you'll only ever get the "on the label" features.