chore(badge): refactor into composable parts
Before this commit, all `badge` code was in a single file.

Now, it's split between three:
* class
* class-extensions
* theme

There were no changes but to the structure of the files.

This is in support of a css-only core, available extensions, and theme-only customization files.

Additionally, documentation was added for previously undocumented APIs, as bad as they might be.