style(checkbox): fix checkmark character
Before this commit, checkbox characters were broken.

This is a result of the shift to InterfacesIcons. Using `content` mimick fonts has long plagued us. Trying to keep them consistent is a HUGE pain in the ass and I don't plan to do so. In most cases we can fix it in markup.

This commit inlines a checkmark in one of the few places where this is not possible.