change(engine): remove auto-injection of Interfaces helpers (#209)
Before this, Interfaces helpers were automatically injected into the host's `application_controller`.

This made it easier for folks to get started without a lot of setup. But we don't setup now apps much anymore and this approach is vulnerable to API changes in Rails.

Now, apps are responsible for mounting Interfaces helpers into `application_controller`. This can be done by adding one line:

```helper helper Interfaces::Engine.helpers```

Additionally, this change is required for compatibility with Rails 5. There is no auto-injecting branch that works with Rails 5.