fix(v1.4.0): make v1.4.0 great again
`react-media-object.js` was removed in [633faf][]. But it was still
being required from `dependencies.js`. The unintended result is that
`rake assets:precompile` fails in an application that is using
`interfaces`. Removing the Sprockets require directive resolves the
issue. 🎉

Along the way I bumped the version, and conslidated the divergent
CHANGELOG strategies that I noticed while comparing refs on GitHub.

[633faf]: https://github.com/ministrycentered/interfaces/commit/633fafb9daf1bdfacf69c3c4c7b19d91fb852c01.